+++ dev/main-react16/src/pages/react16/react16.js
@@ -279,6 +279,7 @@ export default class App extends React.Component {
                   // disable-scopecss
                   // macro
                   // shadowDOM
+                  // disable-patch-request
                 >
                 </micro-app>
               )