+++ src/__tests__/common/util.ts
@@ -1,4 +1,4 @@
-export async function waitFor (cb: CallableFunction) {
+export async function waitFor (cb: CallableFunction): Promise<any> {
   return new Promise(resolve => {
     setTimeout(() => {
       resolve(cb())