+++ src/__tests__/unit/mocks/fetch.ts
@@ -1,4 +1,4 @@
-export function setupMockFetch (data: string, error?: boolean) {
+export function setupMockFetch (data: string, error?: boolean): CallableFunction {
   return function fetchStub () {
     return error
       ? Promise.reject(new Error('network error'))